Renaming KPIEvents to EPIEventTypes
