+++ core-common/src/main/java/org/zalando/nakadi/config/KPIEventTypes.java
